+++ eslint.config.js
@@ -3,107 +3,107 @@ import pluginVue from 'eslint-plugin-vue'
 import globals from 'globals'
 
 export default [
-  // Ignore patterns
-  {
-    ignores: [
-      'dist/**',
-      'node_modules/**',
-      '.storybook/**',
-      'storybook-static/**',
-      'coverage/**'
-    ]
-  },
-
-  // Base JavaScript config
-  js.configs.recommended,
-
-  // Vue 3 recommended config
-  ...pluginVue.configs['flat/recommended'],
-
-  // Config files specific
-  {
-    files: ['*.config.js', 'eslint.config.js'],
-    languageOptions: {
-      globals: {
-        ...globals.node,
-        process: 'readonly',
-        __dirname: 'readonly',
-        module: 'readonly'
-      }
-    }
-  },
-
-  // Custom rules for the project
-  {
-    files: ['**/*.{js,mjs,vue}'],
-    languageOptions: {
-      ecmaVersion: 'latest',
-      sourceType: 'module',
-      globals: {
-        ...globals.browser,
-        ...globals.node,
-        ...globals.es2021
-      }
+    // Ignore patterns
+    {
+        ignores: [
+            'dist/**',
+            'node_modules/**',
+            '.storybook/**',
+            'storybook-static/**',
+            'coverage/**'
+        ]
     },
-    rules: {
-      // Vue specific rules
-      'vue/multi-word-component-names': 'off',
-      'vue/require-default-prop': 'warn',
-      'vue/no-unused-vars': 'error',
-      'vue/no-v-html': 'warn',
-      'vue/html-indent': ['error', 2],
-      'vue/max-attributes-per-line': ['error', {
-        singleline: 5,
-        multiline: 1
-      }],
-      'vue/html-self-closing': ['error', {
-        html: {
-          void: 'never',
-          normal: 'never',
-          component: 'always'
+
+    // Base JavaScript config
+    js.configs.recommended,
+
+    // Vue 3 recommended config
+    ...pluginVue.configs['flat/recommended'],
+
+    // Config files specific
+    {
+        files: ['*.config.js', 'eslint.config.js'],
+        languageOptions: {
+            globals: {
+                ...globals.node,
+                process: 'readonly',
+                __dirname: 'readonly',
+                module: 'readonly'
+            }
         }
-      }],
-      'vue/component-name-in-template-casing': ['error', 'kebab-case'],
-      'vue/singleline-html-element-content-newline': 'off',
-      'vue/attributes-order': 'off',
+    },
 
-      // JavaScript rules
-      'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-      'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-      'no-unused-vars': ['error', {
-        argsIgnorePattern: '^_',
-        varsIgnorePattern: '^_'
-      }],
-      'prefer-const': 'error',
-      'no-var': 'error',
-      'arrow-spacing': 'error',
-      'comma-dangle': ['error', 'never'],
-      'semi': ['error', 'never'],
-      'quotes': ['error', 'single', { avoidEscape: true }],
-      'indent': ['error', 2, { SwitchCase: 1 }],
-      'object-curly-spacing': ['error', 'always'],
-      'array-bracket-spacing': ['error', 'never'],
-      'space-before-function-paren': ['error', {
-        anonymous: 'always',
-        named: 'never',
-        asyncArrow: 'always'
-      }]
-    }
-  },
+    // Custom rules for the project
+    {
+        files: ['**/*.{js,mjs,vue}'],
+        languageOptions: {
+            ecmaVersion: 'latest',
+            sourceType: 'module',
+            globals: {
+                ...globals.browser,
+                ...globals.node,
+                ...globals.es2021
+            }
+        },
+        rules: {
+            // Vue specific rules
+            'vue/multi-word-component-names': 'off',
+            'vue/require-default-prop': 'warn',
+            'vue/no-unused-vars': 'error',
+            'vue/no-v-html': 'warn',
+            'vue/html-indent': ['error', 4],
+            'vue/max-attributes-per-line': ['error', {
+                singleline: 5,
+                multiline: 1
+            }],
+            'vue/html-self-closing': ['error', {
+                html: {
+                    void: 'never',
+                    normal: 'never',
+                    component: 'always'
+                }
+            }],
+            'vue/component-name-in-template-casing': ['error', 'kebab-case'],
+            'vue/singleline-html-element-content-newline': 'off',
+            'vue/attributes-order': 'off',
 
-  // Test files specific config
-  {
-    files: ['**/*.spec.js', '**/*.test.js'],
-    languageOptions: {
-      globals: {
-        ...globals.vitest,
-        describe: 'readonly',
-        it: 'readonly',
-        expect: 'readonly',
-        vi: 'readonly',
-        beforeEach: 'readonly',
-        afterEach: 'readonly'
-      }
+            // JavaScript rules
+            'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+            'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+            'no-unused-vars': ['error', {
+                argsIgnorePattern: '^_',
+                varsIgnorePattern: '^_'
+            }],
+            'prefer-const': 'error',
+            'no-var': 'error',
+            'arrow-spacing': 'error',
+            'comma-dangle': ['error', 'never'],
+            'semi': ['error', 'never'],
+            'quotes': ['error', 'single', { avoidEscape: true }],
+            'indent': ['error', 4, { SwitchCase: 1 }],
+            'object-curly-spacing': ['error', 'always'],
+            'array-bracket-spacing': ['error', 'never'],
+            'space-before-function-paren': ['error', {
+                anonymous: 'always',
+                named: 'never',
+                asyncArrow: 'always'
+            }]
+        }
+    },
+
+    // Test files specific config
+    {
+        files: ['**/*.spec.js', '**/*.test.js'],
+        languageOptions: {
+            globals: {
+                ...globals.vitest,
+                describe: 'readonly',
+                it: 'readonly',
+                expect: 'readonly',
+                vi: 'readonly',
+                beforeEach: 'readonly',
+                afterEach: 'readonly'
+            }
+        }
     }
-  }
 ]
