+++ src/composables/useChartResize.js
@@ -1,4 +1,4 @@
-import { ref, onMounted, onUnmounted } from 'vue'
+import { onMounted, onUnmounted, ref } from 'vue'
 
 /**
  * Debounce function for resize events
@@ -7,15 +7,15 @@ import { ref, onMounted, onUnmounted } from 'vue'
  * @returns {Function} - Debounced function
  */
 function debounce(func, wait = 16) {
-  let timeout
-  return function executedFunction(...args) {
-    const later = () => {
-      clearTimeout(timeout)
-      func(...args)
+    let timeout
+    return function executedFunction(...args) {
+        const later = () => {
+            clearTimeout(timeout)
+            func(...args)
+        }
+        clearTimeout(timeout)
+        timeout = setTimeout(later, wait)
     }
-    clearTimeout(timeout)
-    timeout = setTimeout(later, wait)
-  }
 }
 
 /**
@@ -25,70 +25,70 @@ function debounce(func, wait = 16) {
  * @returns {Object} - {width, height}
  */
 export function useChartResize(containerRef, callback = null) {
-  const width = ref(0)
-  const height = ref(0)
-  let resizeObserver = null
-  let rafId = null
+    const width = ref(0)
+    const height = ref(0)
+    let resizeObserver = null
+    let rafId = null
 
-  const updateDimensions = () => {
-    // Cancel any pending animation frame
-    if (rafId) {
-      cancelAnimationFrame(rafId)
+    const updateDimensions = () => {
+        // Cancel any pending animation frame
+        if (rafId) {
+            cancelAnimationFrame(rafId)
+        }
+
+        // Use requestAnimationFrame for smooth updates
+        rafId = requestAnimationFrame(() => {
+            if (containerRef.value) {
+                const rect = containerRef.value.getBoundingClientRect()
+                width.value = Math.floor(rect.width)
+                height.value = Math.floor(rect.height)
+
+                if (callback) {
+                    callback({ width: width.value, height: height.value })
+                }
+            }
+        })
     }
 
-    // Use requestAnimationFrame for smooth updates
-    rafId = requestAnimationFrame(() => {
-      if (containerRef.value) {
-        const rect = containerRef.value.getBoundingClientRect()
-        width.value = Math.floor(rect.width)
-        height.value = Math.floor(rect.height)
+    const debouncedUpdate = debounce(updateDimensions, 16)
 
-        if (callback) {
-          callback({ width: width.value, height: height.value })
+    onMounted(() => {
+        updateDimensions()
+
+        // Use ResizeObserver for better performance
+        if (typeof ResizeObserver !== 'undefined') {
+            resizeObserver = new ResizeObserver((entries) => {
+                // Use the more efficient contentRect
+                for (const entry of entries) {
+                    if (entry.target === containerRef.value) {
+                        debouncedUpdate()
+                    }
+                }
+            })
+
+            if (containerRef.value) {
+                resizeObserver.observe(containerRef.value)
+            }
+        } else {
+            // Fallback to window resize event
+            window.addEventListener('resize', debouncedUpdate)
         }
-      }
     })
-  }
 
-  const debouncedUpdate = debounce(updateDimensions, 16)
-
-  onMounted(() => {
-    updateDimensions()
-
-    // Use ResizeObserver for better performance
-    if (typeof ResizeObserver !== 'undefined') {
-      resizeObserver = new ResizeObserver((entries) => {
-        // Use the more efficient contentRect
-        for (const entry of entries) {
-          if (entry.target === containerRef.value) {
-            debouncedUpdate()
-          }
+    onUnmounted(() => {
+        if (rafId) {
+            cancelAnimationFrame(rafId)
         }
-      })
+        if (resizeObserver) {
+            resizeObserver.disconnect()
+        } else {
+            window.removeEventListener('resize', debouncedUpdate)
+        }
+    })
 
-      if (containerRef.value) {
-        resizeObserver.observe(containerRef.value)
-      }
-    } else {
-      // Fallback to window resize event
-      window.addEventListener('resize', debouncedUpdate)
+    return {
+        width,
+        height,
+        updateDimensions
     }
-  })
-
-  onUnmounted(() => {
-    if (rafId) {
-      cancelAnimationFrame(rafId)
-    }
-    if (resizeObserver) {
-      resizeObserver.disconnect()
-    } else {
-      window.removeEventListener('resize', debouncedUpdate)
-    }
-  })
-
-  return {
-    width,
-    height,
-    updateDimensions
-  }
 }
