+++ src/stories/BarChart.stories.js
@@ -1,4 +1,4 @@
-import { ref, onMounted } from 'vue'
+import { onMounted, ref } from 'vue'
 import BarChart from '../components/charts/BarChart.vue'
 
 /**
@@ -45,116 +45,116 @@ import BarChart from '../components/charts/BarChart.vue'
  * ```
  */
 export default {
-  title: 'Charts/BarChart',
-  component: BarChart,
-  parameters: {
-    layout: 'padded'
-  },
-  argTypes: {
-    data: {
-      control: 'object',
-      description: 'Chart data including labels and datasets',
-      table: {
-        type: { summary: 'Object' },
-        defaultValue: { summary: '{}' }
-      }
+    title: 'Charts/BarChart',
+    component: BarChart,
+    parameters: {
+        layout: 'padded'
     },
-    options: {
-      control: 'object',
-      description: 'Chart configuration options',
-      table: {
-        type: { summary: 'Object' },
-        defaultValue: { summary: '{}' }
-      }
-    },
-    width: {
-      control: 'number',
-      description: 'Fixed width (only if responsive is false)',
-      table: {
-        type: { summary: 'Number' },
-        defaultValue: { summary: 'null' }
-      }
-    },
-    height: {
-      control: 'number',
-      description: 'Fixed height (only if responsive is false)',
-      table: {
-        type: { summary: 'Number' },
-        defaultValue: { summary: 'null' }
-      }
-    },
-    onBarClick: {
-      action: 'bar-click',
-      description: 'Event emitted when a bar is clicked',
-      table: {
-        type: { summary: 'Function' }
-      }
-    },
-    onLegendToggle: {
-      action: 'legend-toggle',
-      description: 'Event emitted when legend item is toggled',
-      table: {
-        type: { summary: 'Function' }
-      }
+    argTypes: {
+        data: {
+            control: 'object',
+            description: 'Chart data including labels and datasets',
+            table: {
+                type: { summary: 'Object' },
+                defaultValue: { summary: '{}' }
+            }
+        },
+        options: {
+            control: 'object',
+            description: 'Chart configuration options',
+            table: {
+                type: { summary: 'Object' },
+                defaultValue: { summary: '{}' }
+            }
+        },
+        width: {
+            control: 'number',
+            description: 'Fixed width (only if responsive is false)',
+            table: {
+                type: { summary: 'Number' },
+                defaultValue: { summary: 'null' }
+            }
+        },
+        height: {
+            control: 'number',
+            description: 'Fixed height (only if responsive is false)',
+            table: {
+                type: { summary: 'Number' },
+                defaultValue: { summary: 'null' }
+            }
+        },
+        onBarClick: {
+            action: 'bar-click',
+            description: 'Event emitted when a bar is clicked',
+            table: {
+                type: { summary: 'Function' }
+            }
+        },
+        onLegendToggle: {
+            action: 'legend-toggle',
+            description: 'Event emitted when legend item is toggled',
+            table: {
+                type: { summary: 'Function' }
+            }
+        }
     }
-  }
 }
 
 /**
  * Default Story - Basic bar chart with single dataset
  */
 export const Default = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: `
-      <div style="width: 100%; height: 400px; padding: 20px; background: #f9fafb; border-radius: 8px;">
-        <bar-chart v-bind="args" />
-      </div>
-    `
-  }),
-  args: {
-    data: {
-      labels: ['January', 'February', 'March', 'April', 'May', 'June'],
-      datasets: [{
-        label: 'Sales 2024',
-        data: [12, 19, 3, 5, 2, 8],
-        backgroundColor: '#3b82f6',
-        borderColor: '#2563eb',
-        borderWidth: 2
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        legend: {
-          display: true,
-          position: 'top'
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
         },
-        tooltip: {
-          enabled: true
-        }
-      },
-      scales: {
-        x: {
-          display: true,
-          grid: {
-            display: true
-          }
+        template: `
+            <div style="width: 100%; height: 400px; padding: 20px; background: #f9fafb; border-radius: 8px;">
+                <bar-chart v-bind="args"/>
+            </div>
+        `
+    }),
+    args: {
+        data: {
+            labels: ['January', 'February', 'March', 'April', 'May', 'June'],
+            datasets: [{
+                label: 'Sales 2024',
+                data: [12, 19, 3, 5, 2, 8],
+                backgroundColor: '#3b82f6',
+                borderColor: '#2563eb',
+                borderWidth: 2
+            }]
         },
-        y: {
-          display: true,
-          beginAtZero: true,
-          grid: {
-            display: true
-          }
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            plugins: {
+                legend: {
+                    display: true,
+                    position: 'top'
+                },
+                tooltip: {
+                    enabled: true
+                }
+            },
+            scales: {
+                x: {
+                    display: true,
+                    grid: {
+                        display: true
+                    }
+                },
+                y: {
+                    display: true,
+                    beginAtZero: true,
+                    grid: {
+                        display: true
+                    }
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -171,56 +171,56 @@ export const Default = {
  * - ✓ Tooltip shows correct dataset information
  */
 export const MultipleDatasets = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
-      datasets: [
-        {
-          label: 'Sales 2024',
-          data: [12, 19, 3, 5, 2, 8, 9, 14, 11, 6, 10, 15],
-          backgroundColor: '#3b82f6',
-          borderColor: '#2563eb',
-          borderWidth: 2
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
         },
-        {
-          label: 'Sales 2023',
-          data: [7, 11, 5, 8, 3, 7, 6, 9, 8, 4, 7, 10],
-          backgroundColor: '#8b5cf6',
-          borderColor: '#7c3aed',
-          borderWidth: 2
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
+            datasets: [
+                {
+                    label: 'Sales 2024',
+                    data: [12, 19, 3, 5, 2, 8, 9, 14, 11, 6, 10, 15],
+                    backgroundColor: '#3b82f6',
+                    borderColor: '#2563eb',
+                    borderWidth: 2
+                },
+                {
+                    label: 'Sales 2023',
+                    data: [7, 11, 5, 8, 3, 7, 6, 9, 8, 4, 7, 10],
+                    backgroundColor: '#8b5cf6',
+                    borderColor: '#7c3aed',
+                    borderWidth: 2
+                },
+                {
+                    label: 'Sales 2022',
+                    data: [5, 8, 4, 6, 2, 5, 4, 7, 6, 3, 5, 8],
+                    backgroundColor: '#10b981',
+                    borderColor: '#059669',
+                    borderWidth: 2
+                }
+            ]
         },
-        {
-          label: 'Sales 2022',
-          data: [5, 8, 4, 6, 2, 5, 4, 7, 6, 3, 5, 8],
-          backgroundColor: '#10b981',
-          borderColor: '#059669',
-          borderWidth: 2
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            plugins: {
+                legend: {
+                    display: true,
+                    position: 'top'
+                }
+            },
+            scales: {
+                y: {
+                    beginAtZero: true
+                }
+            }
         }
-      ]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        legend: {
-          display: true,
-          position: 'top'
-        }
-      },
-      scales: {
-        y: {
-          beginAtZero: true
-        }
-      }
     }
-  }
 }
 
 /**
@@ -235,42 +235,42 @@ export const MultipleDatasets = {
  * - ✓ Tick marks and labels display correctly
  */
 export const NoGridLines = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['Q1', 'Q2', 'Q3', 'Q4'],
-      datasets: [{
-        label: 'Revenue',
-        data: [45000, 52000, 48000, 67000],
-        backgroundColor: '#f59e0b',
-        borderColor: '#d97706',
-        borderWidth: 2
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      scales: {
-        x: {
-          grid: {
-            display: false
-          }
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
         },
-        y: {
-          beginAtZero: true,
-          grid: {
-            display: false
-          }
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['Q1', 'Q2', 'Q3', 'Q4'],
+            datasets: [{
+                label: 'Revenue',
+                data: [45000, 52000, 48000, 67000],
+                backgroundColor: '#f59e0b',
+                borderColor: '#d97706',
+                borderWidth: 2
+            }]
+        },
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            scales: {
+                x: {
+                    grid: {
+                        display: false
+                    }
+                },
+                y: {
+                    beginAtZero: true,
+                    grid: {
+                        display: false
+                    }
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -285,34 +285,34 @@ export const NoGridLines = {
  * - ✓ Accessibility standards met for color contrast
  */
 export const CustomColors = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['North', 'South', 'East', 'West', 'Central'],
-      datasets: [{
-        label: 'Regional Sales',
-        data: [85, 92, 78, 88, 95],
-        backgroundColor: ['#ef4444', '#f59e0b', '#10b981', '#3b82f6', '#8b5cf6'],
-        borderColor: ['#dc2626', '#d97706', '#059669', '#2563eb', '#7c3aed'],
-        borderWidth: 2
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        legend: {
-          display: false
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
+        },
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['North', 'South', 'East', 'West', 'Central'],
+            datasets: [{
+                label: 'Regional Sales',
+                data: [85, 92, 78, 88, 95],
+                backgroundColor: ['#ef4444', '#f59e0b', '#10b981', '#3b82f6', '#8b5cf6'],
+                borderColor: ['#dc2626', '#d97706', '#059669', '#2563eb', '#7c3aed'],
+                borderWidth: 2
+            }]
+        },
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            plugins: {
+                legend: {
+                    display: false
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -329,86 +329,88 @@ export const CustomColors = {
  * - ✓ No memory leaks on repeated updates
  */
 export const Interactive = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      const chartData = ref({
-        labels: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'],
-        datasets: [{
-          label: 'Daily Views',
-          data: [120, 190, 130, 150, 220, 180, 140],
-          backgroundColor: '#3b82f6',
-          borderColor: '#2563eb',
-          borderWidth: 2
-        }]
-      })
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            const chartData = ref({
+                labels: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'],
+                datasets: [{
+                    label: 'Daily Views',
+                    data: [120, 190, 130, 150, 220, 180, 140],
+                    backgroundColor: '#3b82f6',
+                    borderColor: '#2563eb',
+                    borderWidth: 2
+                }]
+            })
 
-      const events = ref([])
+            const events = ref([])
 
-      const handleBarClick = (data) => {
-        events.value.unshift({
-          label: data.label,
-          value: data.value,
-          dataset: data.datasetIndex
-        })
-        if (events.value.length > 5) events.value.pop()
-      }
+            const handleBarClick = (data) => {
+                events.value.unshift({
+                    label: data.label,
+                    value: data.value,
+                    dataset: data.datasetIndex
+                })
+                if (events.value.length > 5) events.value.pop()
+            }
 
-      const randomizeData = () => {
-        chartData.value = {
-          ...chartData.value,
-          datasets: chartData.value.datasets.map(dataset => ({
-            ...dataset,
-            data: Array.from({ length: 7 }, () => Math.floor(Math.random() * 300) + 50)
-          }))
-        }
-      }
+            const randomizeData = () => {
+                chartData.value = {
+                    ...chartData.value,
+                    datasets: chartData.value.datasets.map(dataset => ({
+                        ...dataset,
+                        data: Array.from({ length: 7 }, () => Math.floor(Math.random() * 300) + 50)
+                    }))
+                }
+            }
 
-      return { chartData, events, handleBarClick, randomizeData, args }
-    },
-    template: `
-      <div style="width: 100%;">
-        <div style="margin-bottom: 16px;">
-          <button
-            @click="randomizeData"
-            style="padding: 8px 16px; background: #3b82f6; color: white; border: none; border-radius: 6px; cursor: pointer; font-size: 14px; font-weight: 500;"
-          >
-            Randomize Data
-          </button>
-        </div>
-        <div style="width: 100%; height: 400px; margin-bottom: 16px;">
-          <bar-chart
-            :data="chartData"
-            :options="args.options"
-            @bar-click="handleBarClick"
-          />
-        </div>
-        <div v-if="events.length > 0" style="padding: 16px; background: #f9fafb; border: 1px solid #e5e7eb; border-radius: 8px;">
-          <h4 style="margin: 0 0 12px 0; font-size: 14px; font-weight: 600; color: #111827;">Recent Events</h4>
-          <div style="display: flex; flex-direction: column; gap: 8px;">
-            <div
-              v-for="(event, index) in events"
-              :key="index"
-              style="padding: 8px 12px; background: white; border: 1px solid #e5e7eb; border-radius: 4px; font-size: 13px; font-family: monospace;"
-            >
-              Clicked: <strong>{{ event.label }}</strong> - Value: <strong>{{ event.value }}</strong>
+            return { chartData, events, handleBarClick, randomizeData, args }
+        },
+        template: `
+            <div style="width: 100%;">
+                <div style="margin-bottom: 16px;">
+                    <button
+                        @click="randomizeData"
+                        style="padding: 8px 16px; background: #3b82f6; color: white; border: none; border-radius: 6px; cursor: pointer; font-size: 14px; font-weight: 500;"
+                    >
+                        Randomize Data
+                    </button>
+                </div>
+                <div style="width: 100%; height: 400px; margin-bottom: 16px;">
+                    <bar-chart
+                        :data="chartData"
+                        :options="args.options"
+                        @bar-click="handleBarClick"
+                    />
+                </div>
+                <div v-if="events.length > 0"
+                     style="padding: 16px; background: #f9fafb; border: 1px solid #e5e7eb; border-radius: 8px;">
+                    <h4 style="margin: 0 0 12px 0; font-size: 14px; font-weight: 600; color: #111827;">Recent
+                        Events</h4>
+                    <div style="display: flex; flex-direction: column; gap: 8px;">
+                        <div
+                            v-for="(event, index) in events"
+                            :key="index"
+                            style="padding: 8px 12px; background: white; border: 1px solid #e5e7eb; border-radius: 4px; font-size: 13px; font-family: monospace;"
+                        >
+                            Clicked: <strong>{{ event.label }}</strong> - Value: <strong>{{ event.value }}</strong>
+                        </div>
+                    </div>
+                </div>
             </div>
-          </div>
-        </div>
-      </div>
-    `
-  }),
-  args: {
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        tooltip: {
-          enabled: true
+        `
+    }),
+    args: {
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            plugins: {
+                tooltip: {
+                    enabled: true
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -423,29 +425,29 @@ export const Interactive = {
  * - ✓ No layout issues with minimal data
  */
 export const SmallData = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['A', 'B', 'C'],
-      datasets: [{
-        label: 'Values',
-        data: [5, 3, 8],
-        backgroundColor: '#06b6d4',
-        borderColor: '#0891b2',
-        borderWidth: 2
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
+        },
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['A', 'B', 'C'],
+            datasets: [{
+                label: 'Values',
+                data: [5, 3, 8],
+                backgroundColor: '#06b6d4',
+                borderColor: '#0891b2',
+                borderWidth: 2
+            }]
+        },
+        options: {
+            responsive: true,
+            maintainAspectRatio: false
+        }
     }
-  }
 }
 
 /**
@@ -461,36 +463,36 @@ export const SmallData = {
  * - ✓ Memory usage is reasonable
  */
 export const LargeDataset = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: Array.from({ length: 50 }, (_, i) => `Day ${i + 1}`),
-      datasets: [{
-        label: 'Daily Revenue',
-        data: Array.from({ length: 50 }, () => Math.floor(Math.random() * 100) + 20),
-        backgroundColor: '#ec4899',
-        borderColor: '#db2777',
-        borderWidth: 1
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      scales: {
-        x: {
-          ticks: {
-            display: false // Hide labels for large dataset
-          }
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
+        },
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: Array.from({ length: 50 }, (_, i) => `Day ${i + 1}`),
+            datasets: [{
+                label: 'Daily Revenue',
+                data: Array.from({ length: 50 }, () => Math.floor(Math.random() * 100) + 20),
+                backgroundColor: '#ec4899',
+                borderColor: '#db2777',
+                borderWidth: 1
+            }]
+        },
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            scales: {
+                x: {
+                    ticks: {
+                        display: false // Hide labels for large dataset
+                    }
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -505,34 +507,34 @@ export const LargeDataset = {
  * - ✓ Grid lines visible if enabled
  */
 export const EmptyState = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'],
-      datasets: [{
-        label: 'No Data',
-        data: [0, 0, 0, 0, 0, 0],
-        backgroundColor: '#6b7280',
-        borderColor: '#4b5563',
-        borderWidth: 2
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      scales: {
-        y: {
-          beginAtZero: true
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
+        },
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'],
+            datasets: [{
+                label: 'No Data',
+                data: [0, 0, 0, 0, 0, 0],
+                backgroundColor: '#6b7280',
+                borderColor: '#4b5563',
+                borderWidth: 2
+            }]
+        },
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            scales: {
+                y: {
+                    beginAtZero: true
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -547,34 +549,34 @@ export const EmptyState = {
  * - ✓ Tooltips show negative values correctly
  */
 export const NegativeValues = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'],
-      datasets: [{
-        label: 'Profit/Loss',
-        data: [12, -5, 8, -3, 15, 7],
-        backgroundColor: '#3b82f6',
-        borderColor: '#2563eb',
-        borderWidth: 2
-      }]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      scales: {
-        y: {
-          beginAtZero: false // Allow negative values
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
+        },
+        template: '<div style="width: 100%; height: 400px;"><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'],
+            datasets: [{
+                label: 'Profit/Loss',
+                data: [12, -5, 8, -3, 15, 7],
+                backgroundColor: '#3b82f6',
+                borderColor: '#2563eb',
+                borderWidth: 2
+            }]
+        },
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            scales: {
+                y: {
+                    beginAtZero: false // Allow negative values
+                }
+            }
         }
-      }
     }
-  }
 }
 
 /**
@@ -589,31 +591,31 @@ export const NegativeValues = {
  * - ✓ Maintains aspect ratio
  */
 export const FixedSize = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      return { args }
-    },
-    template: '<div><bar-chart v-bind="args" /></div>'
-  }),
-  args: {
-    data: {
-      labels: ['Product A', 'Product B', 'Product C', 'Product D'],
-      datasets: [{
-        label: 'Units Sold',
-        data: [45, 67, 32, 89],
-        backgroundColor: '#14b8a6',
-        borderColor: '#0d9488',
-        borderWidth: 2
-      }]
-    },
-    width: 600,
-    height: 400,
-    options: {
-      responsive: false,
-      maintainAspectRatio: true
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            return { args }
+        },
+        template: '<div><bar-chart v-bind="args" /></div>'
+    }),
+    args: {
+        data: {
+            labels: ['Product A', 'Product B', 'Product C', 'Product D'],
+            datasets: [{
+                label: 'Units Sold',
+                data: [45, 67, 32, 89],
+                backgroundColor: '#14b8a6',
+                borderColor: '#0d9488',
+                borderWidth: 2
+            }]
+        },
+        width: 600,
+        height: 400,
+        options: {
+            responsive: false,
+            maintainAspectRatio: true
+        }
     }
-  }
 }
 
 /**
@@ -629,72 +631,72 @@ export const FixedSize = {
  * - ✓ Smooth transitions between loading states
  */
 export const WithLoadingSpinner = {
-  render: (args) => ({
-    components: { BarChart },
-    setup() {
-      const chartData = ref(null)
-      const isLoading = ref(true)
+    render: (args) => ({
+        components: { BarChart },
+        setup() {
+            const chartData = ref(null)
+            const isLoading = ref(true)
 
-      const loadData = () => {
-        isLoading.value = true
-        chartData.value = null
+            const loadData = () => {
+                isLoading.value = true
+                chartData.value = null
 
-        setTimeout(() => {
-          chartData.value = {
-            labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'],
-            datasets: [{
-              label: 'Sales',
-              data: Array.from({ length: 6 }, () => Math.floor(Math.random() * 100) + 20),
-              backgroundColor: '#3b82f6',
-              borderColor: '#2563eb',
-              borderWidth: 2
-            }]
-          }
-          setTimeout(() => {
-            isLoading.value = false
-          }, 300)
-        }, 1500)
-      }
+                setTimeout(() => {
+                    chartData.value = {
+                        labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'],
+                        datasets: [{
+                            label: 'Sales',
+                            data: Array.from({ length: 6 }, () => Math.floor(Math.random() * 100) + 20),
+                            backgroundColor: '#3b82f6',
+                            borderColor: '#2563eb',
+                            borderWidth: 2
+                        }]
+                    }
+                    setTimeout(() => {
+                        isLoading.value = false
+                    }, 300)
+                }, 1500)
+            }
 
-      onMounted(() => {
-        loadData()
-      })
+            onMounted(() => {
+                loadData()
+            })
 
-      return { chartData, isLoading, loadData, args }
-    },
-    template: `
-      <div style="width: 100%;">
-        <div style="margin-bottom: 16px;">
-          <button
-            @click="loadData"
-            style="padding: 8px 16px; background: #3b82f6; color: white; border: none; border-radius: 6px; cursor: pointer; font-size: 14px; font-weight: 500;"
-          >
-            Reload Data
-          </button>
-        </div>
-        <div style="width: 100%; height: 400px; background: #f9fafb; border-radius: 8px; padding: 20px;">
-          <bar-chart
-            v-if="chartData"
-            :data="chartData"
-            :options="args.options"
-            :show-loading="isLoading"
-            loading-message="Loading chart data..."
-            :loading-delay="0"
-          />
-          <div
-            v-else
-            style="display: flex; align-items: center; justify-content: center; height: 100%; color: #6b7280;"
-          >
-            Click "Reload Data" to load chart
-          </div>
-        </div>
-      </div>
-    `
-  }),
-  args: {
-    options: {
-      responsive: true,
-      maintainAspectRatio: false
+            return { chartData, isLoading, loadData, args }
+        },
+        template: `
+            <div style="width: 100%;">
+                <div style="margin-bottom: 16px;">
+                    <button
+                        @click="loadData"
+                        style="padding: 8px 16px; background: #3b82f6; color: white; border: none; border-radius: 6px; cursor: pointer; font-size: 14px; font-weight: 500;"
+                    >
+                        Reload Data
+                    </button>
+                </div>
+                <div style="width: 100%; height: 400px; background: #f9fafb; border-radius: 8px; padding: 20px;">
+                    <bar-chart
+                        v-if="chartData"
+                        :data="chartData"
+                        :options="args.options"
+                        :show-loading="isLoading"
+                        loading-message="Loading chart data..."
+                        :loading-delay="0"
+                    />
+                    <div
+                        v-else
+                        style="display: flex; align-items: center; justify-content: center; height: 100%; color: #6b7280;"
+                    >
+                        Click "Reload Data" to load chart
+                    </div>
+                </div>
+            </div>
+        `
+    }),
+    args: {
+        options: {
+            responsive: true,
+            maintainAspectRatio: false
+        }
     }
-  }
 }
