+++ src/utils/chartCalculations.js
@@ -4,17 +4,17 @@
  * @returns {Object} - {min, max}
  */
 export function getDataRange(datasets) {
-  let min = Infinity
-  let max = -Infinity
+    let min = Infinity
+    let max = -Infinity
 
-  datasets.forEach(dataset => {
-    dataset.data.forEach(value => {
-      if (value < min) min = value
-      if (value > max) max = value
+    datasets.forEach(dataset => {
+        dataset.data.forEach(value => {
+            if (value < min) min = value
+            if (value > max) max = value
+        })
     })
-  })
 
-  return { min, max }
+    return { min, max }
 }
 
 /**
@@ -25,30 +25,30 @@ export function getDataRange(datasets) {
  * @returns {Object} - {min, max, step}
  */
 export function calculateNiceScale(min, max, ticks = 5) {
-  const range = max - min
-  const roughStep = range / (ticks - 1)
-  const magnitude = Math.floor(Math.log10(roughStep))
-  const magnitudePower = Math.pow(10, magnitude)
-  const possibleSteps = [1, 2, 5, 10]
+    const range = max - min
+    const roughStep = range / (ticks - 1)
+    const magnitude = Math.floor(Math.log10(roughStep))
+    const magnitudePower = Math.pow(10, magnitude)
+    const possibleSteps = [1, 2, 5, 10]
 
-  let niceStep = possibleSteps[0] * magnitudePower
-  for (const step of possibleSteps) {
-    const testStep = step * magnitudePower
-    if (testStep >= roughStep) {
-      niceStep = testStep
-      break
+    let niceStep = possibleSteps[0] * magnitudePower
+    for (const step of possibleSteps) {
+        const testStep = step * magnitudePower
+        if (testStep >= roughStep) {
+            niceStep = testStep
+            break
+        }
     }
-  }
 
-  const niceMin = Math.floor(min / niceStep) * niceStep
-  const niceMax = Math.ceil(max / niceStep) * niceStep
+    const niceMin = Math.floor(min / niceStep) * niceStep
+    const niceMax = Math.ceil(max / niceStep) * niceStep
 
-  return {
-    min: niceMin,
-    max: niceMax,
-    step: niceStep,
-    ticks: Math.round((niceMax - niceMin) / niceStep) + 1
-  }
+    return {
+        min: niceMin,
+        max: niceMax,
+        step: niceStep,
+        ticks: Math.round((niceMax - niceMin) / niceStep) + 1
+    }
 }
 
 /**
@@ -57,18 +57,18 @@ export function calculateNiceScale(min, max, ticks = 5) {
  * @returns {Array} - Array of stacked totals
  */
 export function calculateStackedValues(datasets) {
-  if (!datasets || datasets.length === 0) return []
+    if (!datasets || datasets.length === 0) return []
 
-  const length = datasets[0].data.length
-  const stacked = new Array(length).fill(0)
+    const length = datasets[0].data.length
+    const stacked = new Array(length).fill(0)
 
-  datasets.forEach(dataset => {
-    dataset.data.forEach((value, index) => {
-      stacked[index] += value
+    datasets.forEach(dataset => {
+        dataset.data.forEach((value, index) => {
+            stacked[index] += value
+        })
     })
-  })
 
-  return stacked
+    return stacked
 }
 
 /**
@@ -77,13 +77,13 @@ export function calculateStackedValues(datasets) {
  * @returns {Array} - Array of percentages
  */
 export function calculatePercentages(data) {
-  const total = data.reduce((sum, value) => sum + value, 0)
+    const total = data.reduce((sum, value) => sum + value, 0)
 
-  if (total === 0) {
-    return data.map(() => 0)
-  }
+    if (total === 0) {
+        return data.map(() => 0)
+    }
 
-  return data.map(value => (value / total) * 100)
+    return data.map(value => (value / total) * 100)
 }
 
 /**
@@ -92,22 +92,22 @@ export function calculatePercentages(data) {
  * @returns {Array} - Array of {startAngle, endAngle, percentage}
  */
 export function calculatePieSlices(data) {
-  const percentages = calculatePercentages(data)
-  const slices = []
-  let currentAngle = -90 // Start at top
+    const percentages = calculatePercentages(data)
+    const slices = []
+    let currentAngle = -90 // Start at top
 
-  percentages.forEach((percentage, index) => {
-    const angle = (percentage / 100) * 360
-    slices.push({
-      startAngle: currentAngle,
-      endAngle: currentAngle + angle,
-      percentage,
-      value: data[index]
+    percentages.forEach((percentage, index) => {
+        const angle = (percentage / 100) * 360
+        slices.push({
+            startAngle: currentAngle,
+            endAngle: currentAngle + angle,
+            percentage,
+            value: data[index]
+        })
+        currentAngle += angle
     })
-    currentAngle += angle
-  })
 
-  return slices
+    return slices
 }
 
 /**
@@ -119,12 +119,12 @@ export function calculatePieSlices(data) {
  * @returns {Object} - {x, y}
  */
 export function polarToCartesian(centerX, centerY, radius, angleInDegrees) {
-  const angleInRadians = (angleInDegrees - 90) * Math.PI / 180.0
+    const angleInRadians = (angleInDegrees - 90) * Math.PI / 180.0
 
-  return {
-    x: centerX + (radius * Math.cos(angleInRadians)),
-    y: centerY + (radius * Math.sin(angleInRadians))
-  }
+    return {
+        x: centerX + (radius * Math.cos(angleInRadians)),
+        y: centerY + (radius * Math.sin(angleInRadians))
+    }
 }
 
 /**
@@ -137,16 +137,16 @@ export function polarToCartesian(centerX, centerY, radius, angleInDegrees) {
  * @returns {string} - SVG path string
  */
 export function describePieSlice(centerX, centerY, radius, startAngle, endAngle) {
-  const start = polarToCartesian(centerX, centerY, radius, endAngle)
-  const end = polarToCartesian(centerX, centerY, radius, startAngle)
-  const largeArcFlag = endAngle - startAngle <= 180 ? '0' : '1'
+    const start = polarToCartesian(centerX, centerY, radius, endAngle)
+    const end = polarToCartesian(centerX, centerY, radius, startAngle)
+    const largeArcFlag = endAngle - startAngle <= 180 ? '0' : '1'
 
-  return [
-    'M', centerX, centerY,
-    'L', start.x, start.y,
-    'A', radius, radius, 0, largeArcFlag, 0, end.x, end.y,
-    'Z'
-  ].join(' ')
+    return [
+        'M', centerX, centerY,
+        'L', start.x, start.y,
+        'A', radius, radius, 0, largeArcFlag, 0, end.x, end.y,
+        'Z'
+    ].join(' ')
 }
 
 /**
@@ -157,8 +157,8 @@ export function describePieSlice(centerX, centerY, radius, startAngle, endAngle)
  * @returns {number} - Bar width
  */
 export function calculateBarWidth(chartWidth, dataLength, padding = 0.2) {
-  const availableWidth = chartWidth / dataLength
-  return availableWidth * (1 - padding)
+    const availableWidth = chartWidth / dataLength
+    return availableWidth * (1 - padding)
 }
 
 /**
@@ -168,13 +168,13 @@ export function calculateBarWidth(chartWidth, dataLength, padding = 0.2) {
  * @returns {string} - Formatted number
  */
 export function formatNumber(value, decimals = 2) {
-  if (Math.abs(value) >= 1000000) {
-    return (value / 1000000).toFixed(decimals) + 'M'
-  }
-  if (Math.abs(value) >= 1000) {
-    return (value / 1000).toFixed(decimals) + 'K'
-  }
-  return value.toFixed(decimals)
+    if (Math.abs(value) >= 1000000) {
+        return (value / 1000000).toFixed(decimals) + 'M'
+    }
+    if (Math.abs(value) >= 1000) {
+        return (value / 1000).toFixed(decimals) + 'K'
+    }
+    return value.toFixed(decimals)
 }
 
 /**
@@ -185,5 +185,5 @@ export function formatNumber(value, decimals = 2) {
  * @returns {number} - Interpolated value
  */
 export function interpolate(start, end, progress) {
-  return start + (end - start) * progress
+    return start + (end - start) * progress
 }
