+++ src/utils/colourUtils.js
@@ -2,16 +2,16 @@
  * Default color palette for charts
  */
 export const DEFAULT_COLORS = [
-  '#3b82f6', // blue
-  '#8b5cf6', // purple
-  '#10b981', // green
-  '#f59e0b', // orange
-  '#ef4444', // red
-  '#06b6d4', // cyan
-  '#ec4899', // pink
-  '#14b8a6', // teal
-  '#f97316', // orange-red
-  '#a855f7' // violet
+    '#3b82f6', // blue
+    '#8b5cf6', // purple
+    '#10b981', // green
+    '#f59e0b', // orange
+    '#ef4444', // red
+    '#06b6d4', // cyan
+    '#ec4899', // pink
+    '#14b8a6', // teal
+    '#f97316', // orange-red
+    '#a855f7' // violet
 ]
 
 /**
@@ -21,7 +21,7 @@ export const DEFAULT_COLORS = [
  * @returns {string} - Color value
  */
 export function getColorByIndex(index, palette = DEFAULT_COLORS) {
-  return palette[index % palette.length]
+    return palette[index % palette.length]
 }
 
 /**
@@ -31,14 +31,14 @@ export function getColorByIndex(index, palette = DEFAULT_COLORS) {
  * @returns {Array<string>} - Array of colors
  */
 export function generateColorPalette(count, customPalette = null) {
-  const palette = customPalette || DEFAULT_COLORS
-  const colors = []
+    const palette = customPalette || DEFAULT_COLORS
+    const colors = []
 
-  for (let i = 0; i < count; i++) {
-    colors.push(getColorByIndex(i, palette))
-  }
+    for (let i = 0; i < count; i++) {
+        colors.push(getColorByIndex(i, palette))
+    }
 
-  return colors
+    return colors
 }
 
 /**
@@ -47,14 +47,14 @@ export function generateColorPalette(count, customPalette = null) {
  * @returns {Object} - RGB object {r, g, b}
  */
 export function hexToRgb(hex) {
-  const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex)
-  return result
-    ? {
-      r: parseInt(result[1], 16),
-      g: parseInt(result[2], 16),
-      b: parseInt(result[3], 16)
-    }
-    : null
+    const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex)
+    return result
+        ? {
+            r: parseInt(result[1], 16),
+            g: parseInt(result[2], 16),
+            b: parseInt(result[3], 16)
+        }
+        : null
 }
 
 /**
@@ -65,10 +65,10 @@ export function hexToRgb(hex) {
  * @returns {string} - Hex color value
  */
 export function rgbToHex(r, g, b) {
-  return '#' + [r, g, b].map(x => {
-    const hex = x.toString(16)
-    return hex.length === 1 ? '0' + hex : hex
-  }).join('')
+    return '#' + [r, g, b].map(x => {
+        const hex = x.toString(16)
+        return hex.length === 1 ? '0' + hex : hex
+    }).join('')
 }
 
 /**
@@ -78,17 +78,17 @@ export function rgbToHex(r, g, b) {
  * @returns {string} - RGBA color value
  */
 export function addAlpha(color, alpha) {
-  if (color.startsWith('#')) {
-    const rgb = hexToRgb(color)
-    if (!rgb) return color
-    return `rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, ${alpha})`
-  }
+    if (color.startsWith('#')) {
+        const rgb = hexToRgb(color)
+        if (!rgb) return color
+        return `rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, ${alpha})`
+    }
 
-  if (color.startsWith('rgb(')) {
-    return color.replace('rgb(', 'rgba(').replace(')', `, ${alpha})`)
-  }
+    if (color.startsWith('rgb(')) {
+        return color.replace('rgb(', 'rgba(').replace(')', `, ${alpha})`)
+    }
 
-  return color
+    return color
 }
 
 /**
@@ -98,15 +98,15 @@ export function addAlpha(color, alpha) {
  * @returns {string} - Lightened hex color
  */
 export function lightenColor(color, percent) {
-  const rgb = hexToRgb(color)
-  if (!rgb) return color
+    const rgb = hexToRgb(color)
+    if (!rgb) return color
 
-  const amount = Math.round(2.55 * percent)
-  const r = Math.min(255, rgb.r + amount)
-  const g = Math.min(255, rgb.g + amount)
-  const b = Math.min(255, rgb.b + amount)
+    const amount = Math.round(2.55 * percent)
+    const r = Math.min(255, rgb.r + amount)
+    const g = Math.min(255, rgb.g + amount)
+    const b = Math.min(255, rgb.b + amount)
 
-  return rgbToHex(r, g, b)
+    return rgbToHex(r, g, b)
 }
 
 /**
@@ -116,15 +116,15 @@ export function lightenColor(color, percent) {
  * @returns {string} - Darkened hex color
  */
 export function darkenColor(color, percent) {
-  const rgb = hexToRgb(color)
-  if (!rgb) return color
+    const rgb = hexToRgb(color)
+    if (!rgb) return color
 
-  const amount = Math.round(2.55 * percent)
-  const r = Math.max(0, rgb.r - amount)
-  const g = Math.max(0, rgb.g - amount)
-  const b = Math.max(0, rgb.b - amount)
+    const amount = Math.round(2.55 * percent)
+    const r = Math.max(0, rgb.r - amount)
+    const g = Math.max(0, rgb.g - amount)
+    const b = Math.max(0, rgb.b - amount)
 
-  return rgbToHex(r, g, b)
+    return rgbToHex(r, g, b)
 }
 
 /**
@@ -133,11 +133,11 @@ export function darkenColor(color, percent) {
  * @returns {string} - '#000000' or '#ffffff'
  */
 export function getContrastColor(backgroundColor) {
-  const rgb = hexToRgb(backgroundColor)
-  if (!rgb) return '#000000'
+    const rgb = hexToRgb(backgroundColor)
+    if (!rgb) return '#000000'
 
-  // Calculate relative luminance
-  const luminance = (0.299 * rgb.r + 0.587 * rgb.g + 0.114 * rgb.b) / 255
+    // Calculate relative luminance
+    const luminance = (0.299 * rgb.r + 0.587 * rgb.g + 0.114 * rgb.b) / 255
 
-  return luminance > 0.5 ? '#000000' : '#ffffff'
+    return luminance > 0.5 ? '#000000' : '#ffffff'
 }
