+++ src/utils/validators.js
@@ -4,27 +4,27 @@
  * @returns {boolean} - True if valid
  */
 export function validateChartData(data) {
-  if (!data || typeof data !== 'object') {
-    return false
-  }
+    if (!data || typeof data !== 'object') {
+        return false
+    }
 
-  if (!Array.isArray(data.labels)) {
-    return false
-  }
+    if (!Array.isArray(data.labels)) {
+        return false
+    }
 
-  if (!Array.isArray(data.datasets) || data.datasets.length === 0) {
-    return false
-  }
+    if (!Array.isArray(data.datasets) || data.datasets.length === 0) {
+        return false
+    }
 
-  // Validate each dataset
-  return data.datasets.every(dataset => {
-    return (
-      dataset &&
-      typeof dataset === 'object' &&
-      Array.isArray(dataset.data) &&
-      dataset.data.length === data.labels.length
-    )
-  })
+    // Validate each dataset
+    return data.datasets.every(dataset => {
+        return (
+            dataset &&
+            typeof dataset === 'object' &&
+            Array.isArray(dataset.data) &&
+            dataset.data.length === data.labels.length
+        )
+    })
 }
 
 /**
@@ -33,15 +33,15 @@ export function validateChartData(data) {
  * @returns {boolean} - True if valid
  */
 export function validateChartOptions(options) {
-  if (!options) {
-    return true // Options are optional
-  }
+    if (!options) {
+        return true // Options are optional
+    }
 
-  if (typeof options !== 'object') {
-    return false
-  }
+    if (typeof options !== 'object') {
+        return false
+    }
 
-  return true
+    return true
 }
 
 /**
@@ -50,13 +50,13 @@ export function validateChartOptions(options) {
  * @returns {boolean} - True if all values are numbers
  */
 export function validateNumericData(data) {
-  if (!Array.isArray(data)) {
-    return false
-  }
+    if (!Array.isArray(data)) {
+        return false
+    }
 
-  return data.every(value => {
-    return typeof value === 'number' && !isNaN(value)
-  })
+    return data.every(value => {
+        return typeof value === 'number' && !isNaN(value)
+    })
 }
 
 /**
@@ -65,30 +65,30 @@ export function validateNumericData(data) {
  * @returns {boolean} - True if valid color
  */
 export function validateColor(color) {
-  if (!color || typeof color !== 'string') {
+    if (!color || typeof color !== 'string') {
+        return false
+    }
+
+    // Test hex colors
+    if (/^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/.test(color)) {
+        return true
+    }
+
+    // Test rgb/rgba
+    if (/^rgb(a)?\([\d,.\s]+\)$/.test(color)) {
+        return true
+    }
+
+    // Test named colors (basic check)
+    const namedColors = [
+        'black', 'white', 'red', 'green', 'blue', 'yellow', 'orange',
+        'purple', 'pink', 'brown', 'gray', 'grey'
+    ]
+    if (namedColors.includes(color.toLowerCase())) {
+        return true
+    }
+
     return false
-  }
-
-  // Test hex colors
-  if (/^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/.test(color)) {
-    return true
-  }
-
-  // Test rgb/rgba
-  if (/^rgb(a)?\([\d,.\s]+\)$/.test(color)) {
-    return true
-  }
-
-  // Test named colors (basic check)
-  const namedColors = [
-    'black', 'white', 'red', 'green', 'blue', 'yellow', 'orange',
-    'purple', 'pink', 'brown', 'gray', 'grey'
-  ]
-  if (namedColors.includes(color.toLowerCase())) {
-    return true
-  }
-
-  return false
 }
 
 /**
@@ -98,17 +98,17 @@ export function validateColor(color) {
  * @returns {boolean} - True if valid
  */
 export function validateDataset(dataset, expectedLength) {
-  if (!dataset || typeof dataset !== 'object') {
-    return false
-  }
+    if (!dataset || typeof dataset !== 'object') {
+        return false
+    }
 
-  if (!Array.isArray(dataset.data)) {
-    return false
-  }
+    if (!Array.isArray(dataset.data)) {
+        return false
+    }
 
-  if (expectedLength !== undefined && dataset.data.length !== expectedLength) {
-    return false
-  }
+    if (expectedLength !== undefined && dataset.data.length !== expectedLength) {
+        return false
+    }
 
-  return true
+    return true
 }
