+++ vite.config.js
@@ -4,48 +4,48 @@ import vue from '@vitejs/plugin-vue'
 import tailwindcss from '@tailwindcss/vite'
 
 export default defineConfig({
-  plugins: [
-    vue(),
-    tailwindcss()
-  ],
-  build: {
-    lib: {
-      entry: resolve(__dirname, 'src/index.js'),
-      name: 'VueCharts',
-      fileName: (format) => `vue-charts.${format === 'es' ? 'js' : 'umd.cjs'}`
-    },
-    cssCodeSplit: false,
-    rollupOptions: {
-      external: ['vue'],
-      output: {
-        globals: {
-          vue: 'Vue'
+    plugins: [
+        vue(),
+        tailwindcss()
+    ],
+    build: {
+        lib: {
+            entry: resolve(__dirname, 'src/index.js'),
+            name: 'VueCharts',
+            fileName: (format) => `vue-charts.${format === 'es' ? 'js' : 'umd.cjs'}`
         },
-        assetFileNames: (assetInfo) => {
-          if (assetInfo.name === 'style.css') return 'style.css'
-          return assetInfo.name
+        cssCodeSplit: false,
+        rollupOptions: {
+            external: ['vue'],
+            output: {
+                globals: {
+                    vue: 'Vue'
+                },
+                assetFileNames: (assetInfo) => {
+                    if (assetInfo.name === 'style.css') return 'style.css'
+                    return assetInfo.name
+                }
+            }
         }
-      }
+    },
+    resolve: {
+        alias: {
+            '@': resolve(__dirname, 'src')
+        }
+    },
+    test: {
+        globals: true,
+        environment: 'happy-dom',
+        coverage: {
+            provider: 'v8',
+            reporter: ['text', 'json', 'html'],
+            exclude: [
+                'node_modules/',
+                'src/stories/',
+                '*.config.js',
+                'dist/',
+                '.storybook/'
+            ]
+        }
     }
-  },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, 'src')
-    }
-  },
-  test: {
-    globals: true,
-    environment: 'happy-dom',
-    coverage: {
-      provider: 'v8',
-      reporter: ['text', 'json', 'html'],
-      exclude: [
-        'node_modules/',
-        'src/stories/',
-        '*.config.js',
-        'dist/',
-        '.storybook/'
-      ]
-    }
-  }
 })
