+++ vitest.config.js
@@ -3,32 +3,32 @@ import vue from '@vitejs/plugin-vue'
 import { resolve } from 'path'
 
 export default defineConfig({
-  plugins: [vue()],
-  test: {
-    globals: true,
-    environment: 'happy-dom',
-    coverage: {
-      provider: 'v8',
-      reporter: ['text', 'json', 'html', 'lcov'],
-      exclude: [
-        'node_modules/',
-        'src/stories/',
-        '*.config.js',
-        'dist/',
-        '.storybook/',
-        'demo/'
-      ],
-      thresholds: {
-        lines: 90,
-        functions: 90,
-        branches: 90,
-        statements: 90
-      }
+    plugins: [vue()],
+    test: {
+        globals: true,
+        environment: 'happy-dom',
+        coverage: {
+            provider: 'v8',
+            reporter: ['text', 'json', 'html', 'lcov'],
+            exclude: [
+                'node_modules/',
+                'src/stories/',
+                '*.config.js',
+                'dist/',
+                '.storybook/',
+                'demo/'
+            ],
+            thresholds: {
+                lines: 90,
+                functions: 90,
+                branches: 90,
+                statements: 90
+            }
+        }
+    },
+    resolve: {
+        alias: {
+            '@': resolve(__dirname, 'src')
+        }
     }
-  },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, 'src')
-    }
-  }
 })
